added functionality that assings new task date stamp ID and sets completed key to false. Entered input is not being added to list yet.
