+++ src/App.js
@@ -2,17 +2,44 @@ import React from 'react';
 import TodoForm from './components/TodoComponents/TodoForm';
 import TodoList from './components/TodoComponents/TodoList';
 
+const taskItems = [
+  {
+    task: 'Organize Garage',
+    id: 1528817077286,
+    completed: false
+  },
+  {
+    task: 'Bake Cookies',
+    id: 1528817084358,
+    completed: false
+  }
+];
 
 class App extends React.Component {
   // you will need a place to store your state in this component.
   // design `App` to be the parent component of your application.
   // this component is going to take care of state, and any change handlers you need to work with your state
+  constructor() {
+    super();
+    this.state = {
+      todo: taskItems
+    }
+  }
+
+  handleAddTask = (event) => {
+    const taskItems = this.state.todo.slice();
+    console.log(taskItems);
+    taskItems.push({task: event.target.value, id: Date.now(), completed: false});
+    console.log(event.target.value)
+    this.setState({todo: taskItems});
+  }
+  
   render() {
     return (
       <div>
         <h2>Welcome to your Todo App!</h2>
-        <TodoList />
-        <TodoForm />
+        <TodoList items={this.state.todo}/>
+        <TodoForm onItemAdd={this.handleAddTask}/>
       </div>
     );
   }
