+++ src/components/TodoComponents/Todo.js
@@ -1,0 +1,10 @@
+import React from "react";
+
+const ListItemRenderer = props => {
+  const {task} = props.taskItems;
+  return (
+    <p>{task}</p>
+  )
+}
+
+export default ListItemRenderer;
