+++ src/components/TodoComponents/TodoForm.js
@@ -1,14 +1,15 @@
 import React from 'react';
 
-const TodoForm = props => {
-  return (
-    <div class="todoForm">
-      <input placeholder="enter task" />
-      <input type="submit" value="Submit"></input>
-      <button>Add todo</button>
-      <button>Clear completed</button>
-    </div>
-  )
+class TodoForm extends React.Component {
+  render() {
+    return (
+      <div className="todoForm">
+        <input placeholder="enter task" onChange={this.handleInputChange}/>
+        <button onClick={this.props.onItemAdd}>Add todo</button>
+        <button>Clear completed</button>
+      </div>
+    );
+  }
 }
 
 export default TodoForm;
